// dashboard/src/components/office-pixel/types.ts
export interface GatewayEvent {
  type: string;
  data?: any;
}